+++ library/src/main/java/com/getbase/android/db/fluentsqlite/Query.java
@@ -76,6 +76,10 @@ public final class Query {
     Preconditions.checkNotNull(queries);
     Preconditions.checkArgument(queries.length > 0);
 
+    if (queries.length == 1) {
+      return queries[0];
+    }
+
     CompoundQueryBuilder builder = select(queries[0]);
     for (int i = 1; i < queries.length; i++) {
       builder = builder.union(queries[i]);
@@ -88,6 +92,10 @@ public final class Query {
     Preconditions.checkNotNull(queries);
     Preconditions.checkArgument(queries.length > 0);
 
+    if (queries.length == 1) {
+      return queries[0];
+    }
+
     CompoundQueryBuilder builder = select(queries[0]);
     for (int i = 1; i < queries.length; i++) {
       builder = builder.unionAll(queries[i]);
@@ -99,6 +107,10 @@ public final class Query {
   public static Query intersect(Query... queries) {
     Preconditions.checkNotNull(queries);
     Preconditions.checkArgument(queries.length > 0);
+
+    if (queries.length == 1) {
+      return queries[0];
+    }
 
     CompoundQueryBuilder builder = select(queries[0]);
     for (int i = 1; i < queries.length; i++) {
@@ -1030,8 +1042,9 @@ public final class Query {
       }
     }
 
+    @SafeVarargs
     @Override
-    public QueryBuilder where(String selection, Object... selectionArgs) {
+    public final <T> QueryBuilder where(String selection, T... selectionArgs) {
       if (!Strings.isNullOrEmpty(selection)) {
         mCurrentQueryPart.mSelection.add(selection);
         if (selectionArgs != null) {
@@ -1042,8 +1055,9 @@ public final class Query {
       return this;
     }
 
+    @SafeVarargs
     @Override
-    public QueryBuilder where(Expression selection, Object... selectionArgs) {
+    public final <T> QueryBuilder where(Expression selection, T... selectionArgs) {
       if (selection != null) {
         mCurrentQueryPart.mTablesUsedInExpressions.addAll(selection.getTables());
         where(selection.getSql(), selection.getMergedArgs(selectionArgs));
@@ -1263,13 +1277,15 @@ public final class Query {
       return mDelegate.orderBy(expression);
     }
 
+    @SafeVarargs
     @Override
-    public QueryBuilder where(String selection, Object... selectionArgs) {
+    public final <T> QueryBuilder where(String selection, T... selectionArgs) {
       return mDelegate.where(selection, selectionArgs);
     }
 
+    @SafeVarargs
     @Override
-    public QueryBuilder where(Expression selection, Object... selectionArgs) {
+    public final <T> QueryBuilder where(Expression selection, T... selectionArgs) {
       return mDelegate.where(selection, selectionArgs);
     }
 
@@ -1361,9 +1377,10 @@ public final class Query {
     QueryBuilder as(String alias);
   }
 
+  @SuppressWarnings("unchecked")
   public interface SelectionBuilder {
-    QueryBuilder where(String selection, Object... selectionArgs);
-    QueryBuilder where(Expression selection, Object... selectionArgs);
+    <T> QueryBuilder where(String selection, T... selectionArgs);
+    <T> QueryBuilder where(Expression selection, T... selectionArgs);
   }
 
   public interface JoinTypeBuilder extends JoinBuilder {