+++ library/src/main/java/com/getbase/android/db/provider/BackRefBuilder.java
@@ -1,6 +1,6 @@
 package com.getbase.android.db.provider;
 
-import com.google.common.collect.Lists;
+import java.util.Arrays;
 
 public class BackRefBuilder extends BatcherWrapper {
 
@@ -14,12 +14,19 @@ public class BackRefBuilder extends BatcherWrapper {
   }
 
   public BackRefBuilder(BatcherImpl batcher, ConvertibleToOperation... convertible) {
-    this(batcher, Lists.newArrayList(convertible));
+    this(batcher, Arrays.asList(convertible));
   }
 
   public BackRefBuilder withValueBackReference(Insert previousInsert, String columnName) {
     for (ConvertibleToOperation convertible : convertibles) {
-      batcher.putBackRef(convertible, new BackRef(previousInsert, columnName));
+      batcher.putValueBackRef(convertible, new ValueBackRef(previousInsert, columnName));
+    }
+    return this;
+  }
+
+  public BackRefBuilder withSelectionBackReference(Insert previousInsert, int selectionArgumentIndex) {
+    for (ConvertibleToOperation convertible : convertibles) {
+      batcher.putSelectionBackRef(convertible, new SelectionBackRef(previousInsert, selectionArgumentIndex));
     }
     return this;
   }