+++ library/src/main/java/com/getbase/android/db/provider/Batcher.java
@@ -1,7 +1,5 @@
 package com.getbase.android.db.provider;
 
-import com.google.common.collect.Multimap;
-
 import android.content.ContentProvider;
 import android.content.ContentProviderClient;
 import android.content.ContentProviderOperation;
@@ -11,7 +9,6 @@ import android.content.OperationApplicationException;
 import android.os.RemoteException;
 
 import java.util.ArrayList;
-import java.util.List;
 
 public abstract class Batcher {
 
@@ -19,39 +16,43 @@ public abstract class Batcher {
     return new BatcherImpl();
   }
 
-  public abstract Batcher append(Batcher Batcher);
-
   public abstract BackRefBuilder append(ConvertibleToOperation... convertibles);
 
   public abstract BackRefBuilder append(Iterable<ConvertibleToOperation> convertibles);
 
+  public abstract Batcher decorateUrisWith(UriDecorator uriDecorator);
+
   public abstract ArrayList<ContentProviderOperation> operations();
 
-  public ContentProviderResult[] applyBatch(ContentProvider provider) {
+  public final ContentProviderResult[] applyBatch(ContentProvider provider) {
     return applyBatchOrThrow(null, new ContentProviderCrudHandler(provider));
   }
 
-  public ContentProviderResult[] applyBatch(ContentProviderClient providerClient) throws RemoteException, OperationApplicationException {
+  public final ContentProviderResult[] applyBatch(ContentProviderClient providerClient) throws RemoteException, OperationApplicationException {
     return applyBatch(null, new ContentProviderClientCrudHandler(providerClient));
   }
 
-  public ContentProviderResult[] applyBatch(String authority, ContentResolver resolver) throws RemoteException, OperationApplicationException {
+  public final ContentProviderResult[] applyBatch(String authority, ContentResolver resolver) throws RemoteException, OperationApplicationException {
     return applyBatch(authority, new ContentResolverCrudHandler(resolver));
   }
 
-  private ContentProviderResult[] applyBatch(String authority, CrudHandler crudHandler) throws RemoteException, OperationApplicationException {
+  public final ContentProviderResult[] applyBatch(String authority, CrudHandler crudHandler) throws RemoteException, OperationApplicationException {
     return crudHandler.applyBatch(authority, operations());
   }
 
-  public ContentProviderResult[] applyBatchOrThrow(String authority, ContentResolver resolver) {
+  public final ContentProviderResult[] applyBatchOrThrow(String authority, ContentProvider provider) {
+    return applyBatchOrThrow(authority, new ContentProviderCrudHandler(provider));
+  }
+
+  public final ContentProviderResult[] applyBatchOrThrow(String authority, ContentResolver resolver) {
     return applyBatchOrThrow(authority, new ContentResolverCrudHandler(resolver));
   }
 
-  public ContentProviderResult[] applyBatchOrThrow(ContentProviderClient client) {
+  public final ContentProviderResult[] applyBatchOrThrow(ContentProviderClient client) {
     return applyBatchOrThrow(null, new ContentProviderClientCrudHandler(client));
   }
 
-  private ContentProviderResult[] applyBatchOrThrow(String authority, CrudHandler crudHandler) {
+  public final ContentProviderResult[] applyBatchOrThrow(String authority, CrudHandler crudHandler) {
     try {
       return applyBatch(authority, crudHandler);
     } catch (RemoteException | OperationApplicationException e) {
@@ -59,17 +60,23 @@ public abstract class Batcher {
     }
   }
 
-  static class BackRef {
-
+  static class ValueBackRef {
     final Insert parent;
     final String column;
 
-    BackRef(Insert parent, String column) {
+    ValueBackRef(Insert parent, String column) {
       this.parent = parent;
       this.column = column;
     }
   }
 
-  protected abstract Multimap<ConvertibleToOperation, BackRef> getBackRefsMultimap();
-  protected abstract List<ConvertibleToOperation> getConvertibles();
+  static class SelectionBackRef {
+    final Insert parent;
+    final int selectionArgumentIndex;
+
+    SelectionBackRef(Insert parent, int selectionArgumentIndex) {
+      this.parent = parent;
+      this.selectionArgumentIndex = selectionArgumentIndex;
+    }
+  }
 }