+++ library/src/main/java/com/getbase/android/db/provider/BatcherWrapper.java
@@ -1,23 +1,14 @@
 package com.getbase.android.db.provider;
 
-import com.google.common.collect.Multimap;
-
 import android.content.ContentProviderOperation;
 
 import java.util.ArrayList;
-import java.util.List;
 
 class BatcherWrapper extends Batcher {
-
   private final Batcher realBatcher;
 
   BatcherWrapper(Batcher realBatcher) {
     this.realBatcher = realBatcher;
-  }
-
-  @Override
-  public Batcher append(Batcher Batcher) {
-    return realBatcher.append(Batcher);
   }
 
   @Override
@@ -31,17 +22,12 @@ class BatcherWrapper extends Batcher {
   }
 
   @Override
+  public Batcher decorateUrisWith(UriDecorator uriDecorator) {
+    return realBatcher.decorateUrisWith(uriDecorator);
+  }
+
+  @Override
   public ArrayList<ContentProviderOperation> operations() {
     return realBatcher.operations();
-  }
-
-  @Override
-  protected Multimap<ConvertibleToOperation, BackRef> getBackRefsMultimap() {
-    return realBatcher.getBackRefsMultimap();
-  }
-
-  @Override
-  protected List<ConvertibleToOperation> getConvertibles() {
-    return realBatcher.getConvertibles();
   }
 }