+++ library/src/main/java/com/getbase/android/db/provider/Delete.java
@@ -18,7 +18,8 @@ public class Delete extends ProviderAction<Integer> implements ConvertibleToOper
     super(uri);
   }
 
-  public Delete where(String selection, Object... selectionArgs) {
+  @SafeVarargs
+  public final <T> Delete where(String selection, T... selectionArgs) {
     this.selection.append(selection, selectionArgs);
     return this;
   }
@@ -34,13 +35,13 @@ public class Delete extends ProviderAction<Integer> implements ConvertibleToOper
   }
 
   @Override
-  public ContentProviderOperation toContentProviderOperation() {
-    return toContentProviderOperationBuilder().build();
+  public ContentProviderOperation toContentProviderOperation(UriDecorator uriDecorator) {
+    return toContentProviderOperationBuilder(uriDecorator).build();
   }
 
   @Override
-  public Builder toContentProviderOperationBuilder() {
-    return ContentProviderOperation.newDelete(getUri())
+  public Builder toContentProviderOperationBuilder(UriDecorator uriDecorator) {
+    return ContentProviderOperation.newDelete(uriDecorator.decorate(getUri()))
         .withSelection(selection.getSelection(), selection.getSelectionArgs());
   }
 }