+++ changelog.md
@@ -1,6 +1,21 @@
 Change Log
 ==========
 
+Version 0.13.1 *(2015-12-05)*
+----------------------------
+
+ * Fix compilation warnings when using varargs APIs
+
+Version 0.13.0 *(2015-09-23)*
+----------------------------
+
+ * Batcher API changes: removed `Batcher.append(Batcher)`, added `Batcher.decorateUrisWith(UriDecorator)` and `selectionBackRef` support.
+
+Version 0.12.1 *(2015-09-17)*
+----------------------------
+
+ * Another fix for compound queries builder
+
 Version 0.12.0 *(2015-09-07)*
 ----------------------------